refactor: split LLM providers into separate files and migrate Google SDK to google.genai
Extract each provider class (Anthropic, OpenAI, Google, Ollama, Mock) from the
1104-line llm_router.py monolith into individual files under providers/.
Migrate GoogleProvider from the deprecated google.generativeai SDK to the new
google.genai Client API (genai.Client, client.aio.models.generate_content,
types.GenerateContentConfig, types.Part.from_bytes).

Backward-compatible re-exports in llm_router.py ensure existing imports
continue to work unchanged.
