+++ pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
     # LLM providers
     "anthropic>=0.40.0",
     "openai>=1.54.3",
-    "google-generativeai>=0.8.3",
+    "google-genai>=1.0.0",
     "ollama>=0.3.3",
 
     # Document processing (docling is in [pdf] extra — see below)
