+++ src/wikimind/engine/llm_router.py
@@ -7,19 +7,13 @@ Handles selection, fallback, cost tracking, and token budgeting.
 from __future__ import annotations
 
 import asyncio
-import base64
 import json
 import time
 from collections.abc import AsyncIterator
 from dataclasses import dataclass, field
 from typing import Any
 
-import anthropic
-import google.generativeai as genai
-import ollama
-import openai
 import structlog
-from google.generativeai.types import GenerationConfig
 from sqlalchemy import func
 from sqlmodel import select
 
@@ -33,7 +27,7 @@ log = structlog.get_logger()
 
 
 # ---------------------------------------------------------------------------
-# Pricing (USD per 1M tokens) — update as providers change pricing
+# Pricing (USD per 1M tokens) -- update as providers change pricing
 # ---------------------------------------------------------------------------
 
 PRICING = {
@@ -49,22 +43,23 @@ PRICING = {
         "gemini-2.0-flash": {"input": 0.10, "output": 0.40},
     },
     Provider.OLLAMA: {
-        "*": {"input": 0.0, "output": 0.0},  # Free — local
+        "*": {"input": 0.0, "output": 0.0},  # Free -- local
     },
     Provider.MOCK: {
-        "*": {"input": 0.0, "output": 0.0},  # Free — deterministic
+        "*": {"input": 0.0, "output": 0.0},  # Free -- deterministic
     },
 }
 
 
 def _calc_cost(provider: Provider, model: str, input_tokens: int, output_tokens: int) -> float:
+    """Calculate the USD cost of an LLM call based on token counts."""
     provider_pricing = PRICING.get(provider, {})
     model_pricing = provider_pricing.get(model) or provider_pricing.get("*", {"input": 0, "output": 0})
     return (input_tokens * model_pricing["input"] + output_tokens * model_pricing["output"]) / 1_000_000
 
 
 # ---------------------------------------------------------------------------
-# StreamSession — wraps a streaming LLM response
+# StreamSession -- wraps a streaming LLM response
 # ---------------------------------------------------------------------------
 
 
@@ -84,720 +79,21 @@ class StreamSession:
 
 
 # ---------------------------------------------------------------------------
-# Provider implementations
+# Provider imports -- each provider lives in its own file under providers/
 # ---------------------------------------------------------------------------
 
-
-class AnthropicProvider:
-    """Anthropic LLM provider."""
-
-    def __init__(self):
-        api_key = get_api_key("anthropic")
-        if not api_key:
-            raise ValueError("Anthropic API key not configured")
-        self.client = anthropic.AsyncAnthropic(api_key=api_key)
-
-    async def complete(self, request: CompletionRequest, model: str) -> CompletionResponse:
-        """Complete a request using Anthropic."""
-        start = time.monotonic()
-
-        messages = [{"role": m["role"], "content": m["content"]} for m in request.messages]
-
-        response = await self.client.messages.create(
-            model=model,
-            max_tokens=request.max_tokens,
-            temperature=request.temperature,
-            system=request.system,
-            messages=messages,  # type: ignore[arg-type]
-        )
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response.content[0].text  # type: ignore[union-attr]
-        input_tokens = response.usage.input_tokens
-        output_tokens = response.usage.output_tokens
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.ANTHROPIC,
-            model_used=model,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            cost_usd=_calc_cost(Provider.ANTHROPIC, model, input_tokens, output_tokens),
-            latency_ms=latency_ms,
-        )
-
-    async def complete_multimodal(
-        self,
-        system: str,
-        content_parts: list[dict[str, Any]],
-        model: str,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> CompletionResponse:
-        """Complete a multimodal request with text and images using Anthropic.
-
-        Args:
-            system: System prompt.
-            content_parts: List of content blocks. Each block is either
-                ``{"type": "text", "text": "..."}`` or
-                ``{"type": "image", "source": {"type": "base64",
-                "media_type": "image/png", "data": "..."}}``.
-            model: Model name to use.
-            max_tokens: Maximum output tokens.
-            temperature: Sampling temperature.
-
-        Returns:
-            CompletionResponse with the LLM's text output.
-        """
-        start = time.monotonic()
-
-        response = await self.client.messages.create(
-            model=model,
-            max_tokens=max_tokens,
-            temperature=temperature,
-            system=system,
-            messages=[{"role": "user", "content": content_parts}],  # type: ignore[arg-type,typeddict-item]
-        )
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response.content[0].text  # type: ignore[union-attr]
-        input_tokens = response.usage.input_tokens
-        output_tokens = response.usage.output_tokens
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.ANTHROPIC,
-            model_used=model,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            cost_usd=_calc_cost(Provider.ANTHROPIC, model, input_tokens, output_tokens),
-            latency_ms=latency_ms,
-        )
-
-    async def stream(self, request: CompletionRequest, model: str) -> StreamSession:
-        """Stream a completion request using Anthropic."""
-        messages = [{"role": m["role"], "content": m["content"]} for m in request.messages]
-        start = time.monotonic()
-
-        async def _generate() -> AsyncIterator[str]:
-            async with self.client.messages.stream(
-                model=model,
-                max_tokens=request.max_tokens,
-                temperature=request.temperature,
-                system=request.system,
-                messages=messages,  # type: ignore[arg-type]
-            ) as stream_ctx:
-                async for text in stream_ctx.text_stream:
-                    yield text
-
-                final = await stream_ctx.get_final_message()
-                latency_ms = int((time.monotonic() - start) * 1000)
-                input_tokens = final.usage.input_tokens
-                output_tokens = final.usage.output_tokens
-                full_text = "".join(getattr(block, "text", "") for block in final.content)
-                session.result = CompletionResponse(
-                    content=full_text,
-                    provider_used=Provider.ANTHROPIC,
-                    model_used=model,
-                    input_tokens=input_tokens,
-                    output_tokens=output_tokens,
-                    cost_usd=_calc_cost(Provider.ANTHROPIC, model, input_tokens, output_tokens),
-                    latency_ms=latency_ms,
-                )
-
-        session = StreamSession(_chunks=_generate())
-        return session
-
-
-class OpenAIProvider:
-    """OpenAI LLM provider."""
-
-    def __init__(self):
-        api_key = get_api_key("openai")
-        if not api_key:
-            raise ValueError("OpenAI API key not configured")
-        self.client = openai.AsyncOpenAI(api_key=api_key)
-
-    async def complete(self, request: CompletionRequest, model: str) -> CompletionResponse:
-        """Complete a request using OpenAI."""
-        start = time.monotonic()
-
-        messages = [{"role": "system", "content": request.system}]
-        messages.extend(request.messages)
-
-        kwargs: dict[str, object] = dict(
-            model=model,
-            max_tokens=request.max_tokens,
-            temperature=request.temperature,
-            messages=messages,
-        )
-        if request.response_format == "json":
-            kwargs["response_format"] = {"type": "json_object"}
-
-        response = await self.client.chat.completions.create(**kwargs)  # type: ignore[call-overload]
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response.choices[0].message.content
-        input_tokens = response.usage.prompt_tokens
-        output_tokens = response.usage.completion_tokens
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.OPENAI,
-            model_used=model,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            cost_usd=_calc_cost(Provider.OPENAI, model, input_tokens, output_tokens),
-            latency_ms=latency_ms,
-        )
-
-    async def complete_multimodal(
-        self,
-        system: str,
-        content_parts: list[dict[str, Any]],
-        model: str,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> CompletionResponse:
-        """Complete a multimodal request with text and images using OpenAI.
-
-        Translates the Anthropic-style content blocks to OpenAI's
-        ``image_url`` format before calling the API.
-
-        Args:
-            system: System prompt.
-            content_parts: List of content blocks in Anthropic format.
-            model: Model name to use.
-            max_tokens: Maximum output tokens.
-            temperature: Sampling temperature.
-
-        Returns:
-            CompletionResponse with the LLM's text output.
-        """
-        start = time.monotonic()
-
-        # Translate Anthropic content blocks to OpenAI format
-        openai_parts: list[dict[str, Any]] = []
-        for part in content_parts:
-            if part["type"] == "text":
-                openai_parts.append({"type": "text", "text": part["text"]})
-            elif part["type"] == "image":
-                media_type = part["source"]["media_type"]
-                data = part["source"]["data"]
-                openai_parts.append(
-                    {
-                        "type": "image_url",
-                        "image_url": {"url": f"data:{media_type};base64,{data}"},
-                    }
-                )
-
-        messages: list[dict[str, Any]] = [
-            {"role": "system", "content": system},
-            {"role": "user", "content": openai_parts},
-        ]
-
-        response = await self.client.chat.completions.create(
-            model=model,
-            max_tokens=max_tokens,
-            temperature=temperature,
-            messages=messages,  # type: ignore[arg-type]
-        )
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response.choices[0].message.content
-        input_tokens = response.usage.prompt_tokens if response.usage else 0
-        output_tokens = response.usage.completion_tokens if response.usage else 0
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.OPENAI,
-            model_used=model,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            cost_usd=_calc_cost(Provider.OPENAI, model, input_tokens, output_tokens),
-            latency_ms=latency_ms,
-        )
-
-    async def stream(self, request: CompletionRequest, model: str) -> StreamSession:
-        """Stream a completion request using OpenAI."""
-        messages: list[dict[str, str]] = [{"role": "system", "content": request.system}]
-        messages.extend(request.messages)
-        start = time.monotonic()
-
-        kwargs: dict[str, object] = dict(
-            model=model,
-            max_tokens=request.max_tokens,
-            temperature=request.temperature,
-            messages=messages,
-            stream=True,
-            stream_options={"include_usage": True},
-        )
-        if request.response_format == "json":
-            kwargs["response_format"] = {"type": "json_object"}
-
-        async def _generate() -> AsyncIterator[str]:
-            full_text_parts: list[str] = []
-            input_tokens = 0
-            output_tokens = 0
-            response_stream = await self.client.chat.completions.create(
-                **kwargs  # type: ignore[call-overload]
-            )
-            async for chunk in response_stream:  # type: ignore[union-attr]
-                if chunk.usage:
-                    input_tokens = chunk.usage.prompt_tokens or 0
-                    output_tokens = chunk.usage.completion_tokens or 0
-                if chunk.choices and chunk.choices[0].delta.content:
-                    text = chunk.choices[0].delta.content
-                    full_text_parts.append(text)
-                    yield text
-
-            latency_ms = int((time.monotonic() - start) * 1000)
-            session.result = CompletionResponse(
-                content="".join(full_text_parts),
-                provider_used=Provider.OPENAI,
-                model_used=model,
-                input_tokens=input_tokens,
-                output_tokens=output_tokens,
-                cost_usd=_calc_cost(Provider.OPENAI, model, input_tokens, output_tokens),
-                latency_ms=latency_ms,
-            )
-
-        session = StreamSession(_chunks=_generate())
-        return session
-
-
-class OllamaProvider:
-    """Ollama local LLM provider."""
-
-    def __init__(self, base_url: str):
-        self.client = ollama.AsyncClient(host=base_url)
-
-    async def complete(self, request: CompletionRequest, model: str) -> CompletionResponse:
-        """Complete a request using Ollama."""
-        start = time.monotonic()
-
-        messages = [{"role": "system", "content": request.system}]
-        messages.extend(request.messages)
-
-        response = await self.client.chat(
-            model=model,
-            messages=messages,
-            options={"temperature": request.temperature},
-        )
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response["message"]["content"]
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.OLLAMA,
-            model_used=model,
-            input_tokens=0,
-            output_tokens=0,
-            cost_usd=0.0,
-            latency_ms=latency_ms,
-        )
-
-    async def complete_multimodal(
-        self,
-        system: str,
-        content_parts: list[dict[str, Any]],
-        model: str,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> CompletionResponse:
-        """Complete a multimodal request using Ollama.
-
-        Ollama accepts images as base64 strings in the ``images`` field
-        of a message.
-
-        Args:
-            system: System prompt.
-            content_parts: List of content blocks in Anthropic format.
-            model: Model name to use.
-            max_tokens: Maximum output tokens.
-            temperature: Sampling temperature.
-
-        Returns:
-            CompletionResponse with the LLM's text output.
-        """
-        start = time.monotonic()
-
-        # Extract text and images from content parts
-        text_parts: list[str] = []
-        images: list[str] = []
-        for part in content_parts:
-            if part["type"] == "text":
-                text_parts.append(part["text"])
-            elif part["type"] == "image":
-                images.append(part["source"]["data"])
-
-        user_content = "\n".join(text_parts) if text_parts else "Describe these images."
-        messages: list[dict[str, Any]] = [
-            {"role": "system", "content": system},
-            {"role": "user", "content": user_content, "images": images},
-        ]
-
-        response = await self.client.chat(
-            model=model,
-            messages=messages,  # type: ignore[arg-type]
-            options={"temperature": temperature},
-        )
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response["message"]["content"]
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.OLLAMA,
-            model_used=model,
-            input_tokens=0,
-            output_tokens=0,
-            cost_usd=0.0,
-            latency_ms=latency_ms,
-        )
-
-    async def stream(self, request: CompletionRequest, model: str) -> StreamSession:
-        """Stream a completion request using Ollama."""
-        messages: list[dict[str, str]] = [{"role": "system", "content": request.system}]
-        messages.extend(request.messages)
-        start = time.monotonic()
-
-        async def _generate() -> AsyncIterator[str]:
-            full_text_parts: list[str] = []
-            response_stream = await self.client.chat(
-                model=model,
-                messages=messages,
-                options={"temperature": request.temperature},
-                stream=True,
-            )
-            async for chunk in response_stream:  # type: ignore[union-attr]
-                text = chunk["message"]["content"]
-                if text:
-                    full_text_parts.append(text)
-                    yield text
-
-            latency_ms = int((time.monotonic() - start) * 1000)
-            session.result = CompletionResponse(
-                content="".join(full_text_parts),
-                provider_used=Provider.OLLAMA,
-                model_used=model,
-                input_tokens=0,
-                output_tokens=0,
-                cost_usd=0.0,
-                latency_ms=latency_ms,
-            )
-
-        session = StreamSession(_chunks=_generate())
-        return session
-
-
-class GoogleProvider:
-    """Google Gemini LLM provider."""
-
-    def __init__(self) -> None:
-        api_key = get_api_key("google")
-        if not api_key:
-            raise ValueError("Google API key not configured")
-        genai.configure(api_key=api_key)
-
-    async def complete(self, request: CompletionRequest, model: str) -> CompletionResponse:
-        """Complete a request using Google Gemini."""
-        start = time.monotonic()
-
-        gen_model = genai.GenerativeModel(model, system_instruction=request.system)
-
-        contents = [{"role": m["role"], "parts": [m["content"]]} for m in request.messages]
-
-        gen_cfg_kwargs: dict[str, Any] = {
-            "max_output_tokens": request.max_tokens,
-            "temperature": request.temperature,
-        }
-        if request.response_format == "json":
-            gen_cfg_kwargs["response_mime_type"] = "application/json"
-
-        response = await gen_model.generate_content_async(
-            contents,
-            generation_config=GenerationConfig(**gen_cfg_kwargs),
-        )
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response.text
-        usage = response.usage_metadata
-        input_tokens = usage.prompt_token_count if usage else 0
-        output_tokens = usage.candidates_token_count if usage else 0
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.GOOGLE,
-            model_used=model,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            cost_usd=_calc_cost(Provider.GOOGLE, model, input_tokens, output_tokens),
-            latency_ms=latency_ms,
-        )
-
-    async def complete_multimodal(
-        self,
-        system: str,
-        content_parts: list[dict[str, Any]],
-        model: str,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> CompletionResponse:
-        """Complete a multimodal request with text and images using Google Gemini.
-
-        Translates the Anthropic-style content blocks to Google's inline
-        data format before calling the API.
-
-        Args:
-            system: System prompt.
-            content_parts: List of content blocks in Anthropic format.
-            model: Model name to use.
-            max_tokens: Maximum output tokens.
-            temperature: Sampling temperature.
-
-        Returns:
-            CompletionResponse with the LLM's text output.
-        """
-        start = time.monotonic()
-
-        gen_model = genai.GenerativeModel(model, system_instruction=system)
-
-        # Translate Anthropic content blocks to Google format
-        google_parts: list[str | dict[str, str | bytes]] = []
-        for part in content_parts:
-            if part["type"] == "text":
-                google_parts.append(part["text"])
-            elif part["type"] == "image":
-                media_type = part["source"]["media_type"]
-                data_b64 = part["source"]["data"]
-                google_parts.append(
-                    {
-                        "mime_type": media_type,
-                        "data": base64.b64decode(data_b64),
-                    }
-                )
-
-        response = await gen_model.generate_content_async(
-            google_parts,
-            generation_config=GenerationConfig(
-                max_output_tokens=max_tokens,
-                temperature=temperature,
-            ),
-        )
-
-        latency_ms = int((time.monotonic() - start) * 1000)
-        content = response.text
-        usage = response.usage_metadata
-        input_tokens = usage.prompt_token_count if usage else 0
-        output_tokens = usage.candidates_token_count if usage else 0
-
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.GOOGLE,
-            model_used=model,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            cost_usd=_calc_cost(Provider.GOOGLE, model, input_tokens, output_tokens),
-            latency_ms=latency_ms,
-        )
-
-    async def stream(self, request: CompletionRequest, model: str) -> StreamSession:
-        """Stream a completion request using Google Gemini."""
-        start = time.monotonic()
-
-        gen_model = genai.GenerativeModel(model, system_instruction=request.system)
-
-        contents = [{"role": m["role"], "parts": [m["content"]]} for m in request.messages]
-
-        gen_cfg_kwargs: dict[str, Any] = {
-            "max_output_tokens": request.max_tokens,
-            "temperature": request.temperature,
-        }
-        if request.response_format == "json":
-            gen_cfg_kwargs["response_mime_type"] = "application/json"
-
-        async def _generate() -> AsyncIterator[str]:
-            full_text_parts: list[str] = []
-            input_tokens = 0
-            output_tokens = 0
-
-            response = await gen_model.generate_content_async(
-                contents,
-                generation_config=GenerationConfig(**gen_cfg_kwargs),
-                stream=True,
-            )
-            async for chunk in response:
-                text = chunk.text
-                if text:
-                    full_text_parts.append(text)
-                    yield text
-                if chunk.usage_metadata:
-                    if chunk.usage_metadata.prompt_token_count:
-                        input_tokens = chunk.usage_metadata.prompt_token_count
-                    if chunk.usage_metadata.candidates_token_count:
-                        output_tokens = chunk.usage_metadata.candidates_token_count
-
-            latency_ms = int((time.monotonic() - start) * 1000)
-            session.result = CompletionResponse(
-                content="".join(full_text_parts),
-                provider_used=Provider.GOOGLE,
-                model_used=model,
-                input_tokens=input_tokens,
-                output_tokens=output_tokens,
-                cost_usd=_calc_cost(Provider.GOOGLE, model, input_tokens, output_tokens),
-                latency_ms=latency_ms,
-            )
-
-        session = StreamSession(_chunks=_generate())
-        return session
-
-
-class MockProvider:
-    """Deterministic mock provider for CI e2e testing.
-
-    Returns canned JSON keyed off the TaskType so callers that expect
-    CompilationResult / QueryResult shapes can parse the response
-    without a real LLM. Zero cost, zero network, fully deterministic.
-
-    Must be explicitly enabled via ``WIKIMIND_LLM__MOCK__ENABLED=true``
-    AND set as the default provider via ``WIKIMIND_LLM__DEFAULT_PROVIDER=mock``
-    to be selected — disabled by default so it cannot silently
-    intercept real traffic.
-    """
-
-    def __init__(self) -> None:
-        # No config needed; all responses are canned
-        pass
-
-    async def complete(self, request: CompletionRequest, model: str) -> CompletionResponse:
-        """Return a deterministic canned response matching the request's task type."""
-        start = time.monotonic()
-        content = self._response_for(request)
-        latency_ms = int((time.monotonic() - start) * 1000)
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.MOCK,
-            model_used=model,
-            input_tokens=0,
-            output_tokens=0,
-            cost_usd=0.0,
-            latency_ms=latency_ms,
-        )
-
-    async def complete_multimodal(
-        self,
-        system: str,
-        content_parts: list[dict[str, Any]],
-        model: str,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> CompletionResponse:
-        """Return a deterministic description for each image in the request."""
-        start = time.monotonic()
-        # Count images and return one description per image
-        image_count = sum(1 for p in content_parts if p.get("type") == "image")
-        descriptions = [
-            f"[Page {i + 1} description: A visual slide with diagrams and minimal text.]" for i in range(image_count)
-        ]
-        content = "\n\n".join(descriptions) if descriptions else "No images provided."
-        latency_ms = int((time.monotonic() - start) * 1000)
-        return CompletionResponse(
-            content=content,
-            provider_used=Provider.MOCK,
-            model_used=model,
-            input_tokens=0,
-            output_tokens=0,
-            cost_usd=0.0,
-            latency_ms=latency_ms,
-        )
-
-    async def stream(self, request: CompletionRequest, model: str) -> StreamSession:
-        """Stream a deterministic canned response in small chunks."""
-        content = self._response_for(request)
-        start = time.monotonic()
-        chunk_size = 20
-
-        async def _generate() -> AsyncIterator[str]:
-            for i in range(0, len(content), chunk_size):
-                yield content[i : i + chunk_size]
-
-            latency_ms = int((time.monotonic() - start) * 1000)
-            session.result = CompletionResponse(
-                content=content,
-                provider_used=Provider.MOCK,
-                model_used=model,
-                input_tokens=0,
-                output_tokens=0,
-                cost_usd=0.0,
-                latency_ms=latency_ms,
-            )
-
-        session = StreamSession(_chunks=_generate())
-        return session
-
-    @staticmethod
-    def _response_for(request: CompletionRequest) -> str:
-        """Return a canned response body matching the task's expected shape."""
-        if request.task_type == TaskType.COMPILE:
-            return json.dumps(_MOCK_COMPILE_RESPONSE)
-        if request.task_type == TaskType.QA:
-            return json.dumps(_MOCK_QA_RESPONSE)
-        if request.task_type == TaskType.LINT:
-            return json.dumps(_MOCK_LINT_RESPONSE)
-        # Unknown task type: return an empty JSON object so parse_json_response
-        # doesn't crash. Tests that need specific shapes should add a mock
-        # for their task type.
-        return "{}"
-
-
-# Canned responses used by MockProvider. Defined at module level so tests
-# can import and assert against them directly.
-_MOCK_COMPILE_RESPONSE: dict = {
-    "title": "Mock Article",
-    "summary": "A deterministic summary produced by the mock LLM provider for testing.",
-    "key_claims": [
-        {
-            "claim": "This article was produced by the mock LLM provider.",
-            "confidence": "sourced",
-            "quote": "mock provider",
-        }
-    ],
-    "concepts": ["testing", "mock"],
-    "backlink_suggestions": [],
-    "open_questions": ["What is real?"],
-    "article_body": (
-        "## Mock Article\n\n"
-        "This article was produced by the mock LLM provider for deterministic "
-        "e2e testing.\n\n"
-        "## Details\n\n"
-        "The mock provider returns canned responses regardless of input, "
-        "enabling CI to run the full Ask loop without a real LLM API."
-    ),
-}
-
-_MOCK_QA_RESPONSE: dict = {
-    "answer": (
-        "This is a mock answer from the WikiMind mock LLM provider. "
-        "Your question was received and processed deterministically "
-        "for testing purposes."
-    ),
-    "confidence": "high",
-    "sources": ["Mock Article"],
-    "related_articles": [],
-    "new_article_suggested": None,
-    "follow_up_questions": [],
-}
-
-_MOCK_LINT_RESPONSE: dict = {
-    "contradictions": [],
-    "stale_claims": [],
-    "orphan_articles": [],
-    "missing_pages": [],
-    "data_gaps": [],
-}
-
+from wikimind.engine.providers import (  # noqa: E402
+    AnthropicProvider,
+    GoogleProvider,
+    MockProvider,
+    OllamaProvider,
+    OpenAIProvider,
+)
+from wikimind.engine.providers.mock import (  # noqa: E402, F401 -- backward compat re-exports
+    _MOCK_COMPILE_RESPONSE,
+    _MOCK_LINT_RESPONSE,
+    _MOCK_QA_RESPONSE,
+)
 
 # ---------------------------------------------------------------------------
 # Router
@@ -834,6 +130,7 @@ class LLMRouter:
         return order
 
     def _is_provider_available(self, provider: Provider) -> bool:
+        """Check if a provider is enabled and has credentials."""
         cfg = getattr(self.settings.llm, provider.value, None)
         if not cfg or not cfg.enabled:
             return False
@@ -842,6 +139,7 @@ class LLMRouter:
         return bool(get_api_key(provider.value))
 
     async def _get_provider_instance(self, provider: Provider):
+        """Create and return a provider instance."""
         if provider == Provider.ANTHROPIC:
             return AnthropicProvider()
         elif provider == Provider.OPENAI:
@@ -856,10 +154,12 @@ class LLMRouter:
             raise ValueError(f"Provider {provider} not implemented yet")
 
     def _get_model(self, provider: Provider) -> str:
+        """Return the configured model name for a provider."""
         cfg = getattr(self.settings.llm, provider.value, None)
         return cfg.model if cfg else "unknown"
 
     async def _check_budget(self) -> None:
+        """Check monthly budget and emit warnings if thresholds are exceeded."""
         if self._budget_warning_sent and self._budget_exceeded_sent:
             return
